api/requirements.txt
// api/requirements.txt
azure-functions
azure-ai-projects~=1.0.0b11
azure-identity~=1.20.0
ansible-core~=2.17.0
flask


